backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
qdrant-client
google-genai # For Gemini API